Actulizacion del service y controller

+++ src/main/java/com/finanzas/backend/service/PersonaService.java
@@ -1,5 +1,7 @@
 package com.finanzas.backend.service;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -8,8 +10,14 @@ import com.finanzas.backend.repository.PersonaRepository;
 
 @Service
 public class PersonaService {
+
+    private final PersonaRepository personaRepository;
+    
     @Autowired
-    private PersonaRepository personaRepository;
+    public PersonaService(PersonaRepository personaRepository) {
+        this.personaRepository = personaRepository;
+    }
+    
 
     public Persona getPersonaById(Long id) {
         return personaRepository.findById(id).orElse(null);
@@ -28,11 +36,19 @@ public class PersonaService {
             persona.setCorreoElectronico(personaDetails.getCorreoElectronico());
             return personaRepository.save(persona);
         }
-        return null;
+        else {
+            throw new IllegalArgumentException("No se encontró la persona con el ID proporcionado: " + id);
+        }
     }
 
     public void deletePersona(Long id) {
-        personaRepository.deleteById(id);
+        Optional<Persona> optionaPersona = personaRepository.findById(id);
+
+        if (optionaPersona.isPresent()) {
+        	personaRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("No se encontró la persona con el ID proporcionado: " + id);
+        }
     }
 }
